+++ ChangeLog.md
@@ -1,8 +1,14 @@
 # Changelog for epi-sim
 
+## 0.1.2.0
+
+- Implement a visualisation of the infection network in `vis/infection-tree.vg.json`.
+- Implement `Epidemic.Utility.eventsAsJsonTree`.
+- Move the `Epidemic.Simulation` module to `Epidemic.Utility`.
+
 ## 0.1.1.0
 
-- Writ the output to CSV using `cassava`
+- Write the output to CSV using `cassava`
 
 ## 0.1.0.0